Set VERSION in Astro::FITS::HeaderCollection
The Astro::FITS::HeaderCollection namespace is used for tied
subheaders. It's embedded in Header.pm and did not have its
own version number. This meant that Module::Build failed to
write a version into META.yml despite writing an entry for
the namespace in META.yml. This then confused the CPAN indexer
which got upset when an undef version number did not match
the prior CPAN version number.